Changed datetime value received from API from Int to Double

+++ GlucoseViewer/GlucoseViewerApp.swift
@@ -92,7 +92,7 @@ struct APIBgs : Codable {
     var sgv: String
     var trend: Int
     var direction: String
-    var datetime: Int
+    var datetime: Double
     var bgdelta: Int?
 }
